Second commit

+++ src/neironweb/NeironWeb.java
@@ -8,7 +8,10 @@ public class NeironWeb {
     ArrayList<ArrayList<Neiron>> web = new ArrayList<ArrayList<Neiron>>();
   
     public static void main(String[] args) {
-        int[]n = {12,56,23,45};
+        
+        
+        
+        double []n = {12,56,23,45};
         int [] ln = {3,2};
    Web w = new Web(n, 2,ln );
    w.start();
